phidata/stock_analysis/requirements.txt
duckduckgo-search
fastapi
groq
phidata
python-dotenv
sqlalchemy
uvicorn
yfinance
